+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggdag (development version)
 
+* Fixed `tidy_dagitty()` error when processing DAGs with no edges (#159)
+* Fixed `dag_paths()` error when no open paths exist between nodes (#180)
 * Introduced new functions `query_conditional_independence()`, `test_conditional_independence()`, and `ggdag_conditional_independence()` for detecting, testing, and visualizing implied conditional independencies in a given DAG and dataset (#139)
 * Introduced new functions `aes_dag()` and `geom_dag()` to simplify specification of ggplot code for most DAGs. Also refactored most quick plots to use these functions (#121)
 * Added new function `geom_label_repel2()` for more opinionated repelled labels that often look better on DAGs (#132)